pyproject.toml
[project]
name = "notes-mcp"
version = "0.1.0"
description = "MCP server for searching, navigating, and maintaining a Markdown knowledge base"
readme = "README.md"
authors = [
    { name = "Jamie Kirkpatrick", email = "jkp@kirkconsulting.co.uk" }
]
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "structlog>=24.0",
    "pyyaml>=6.0",
]

[project.scripts]
notes-mcp = "notes_mcp.server:main"

[build-system]
requires = ["uv_build>=0.9.17,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25",
    "pytest-cov>=6.0",
    "mypy>=1.0",
    "ruff>=0.9",
    "pytest-timeout>=2.3",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["E501"]

[tool.mypy]
python_version = "3.13"
strict = true
